+++ History.md
@@ -1,3 +1,7 @@
+## [2.7.6](https://github.com/cucumber/gherkin/compare/v2.7.5...v2.7.6)
+
+* [Java] Previous release accidentally had pretty formatting for both JSONFormatter and JSONPrettyFormatter (Aslak Hellesøy)
+
 ## [2.7.5](https://github.com/cucumber/gherkin/compare/v2.7.4...v2.7.5)
 
 * [Java] Added Formatter.close() so we can close underlying streams after a call to Formatter.done() (Aslak Hellesøy)